Fix circle ci config
This config wasn't getting used at all, the directory name was
incorrect. Tweak config slightly based off of the docs and remove Heroku
config for now, we're using Heroku's deployement infrastructure instead.

This also runs system tests, we were skipping those before. I'm
hard-coding the chromedriver version to avoid a conflict with the chrome
version, we may need to rethink this in the future.